+++ CHANGELOG.md
@@ -35,6 +35,9 @@
   queue for pinging session and that message relies on existing implicit subscription already present
   that would be removed as an orphan subscription.
 * Fix bug in the `AbstractSessionRestService` so that the `synchronized` flag is correct. 
+* Fix bug in `Converger.convergeAreaOfInterest(...)` so that if an `AreaOfInterest` is added when an
+  existing `Subscription` already exists, the `AreaOfInterest` is marked as `LOADED` rather than getting
+  stuck in the `NOT_ASKED` state.
 
 ### [v6.02](https://github.com/realityforge/replicant/tree/v6.02) (2018-07-03)
 [Full Changelog](https://github.com/realityforge/replicant/compare/v6.01...v6.02)